+++ plugins/moxie/src/utils/index.ts
@@ -1,3 +1,4 @@
 export * from './escape'
 export * from './toKebab'
 export * from './transformerUtils'
+export * from './processorUtils'